infrastructure/pyproject.toml
[project]
name = "infrastructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-aws-bedrock-agentcore-alpha>=2.232.1a0",
    "aws-cdk-lib>=2.232.1",
    "constructs>=10.4.3",
]
